fix: bon earnings query duplicate column error when no earnings exist
A new column is selected for every earning, but if there are no earnings, then no columns are selected, and laravel selects `*` by default. Change it to select `1` by default instead, so that it doesn't fall back to `*` when no earnings exist. This selected column won't ever be used, but won't cause any errors either.