fix: search by tmdb 0 to find missing tmdbs
Not sure what was going through my mind when deciding the query for this the first time around, it didn't exactly do anything useful. Now it's back to doing what it used to do before the switch to nulls instead of 0s for metadata ids: finding torrents without any associated tmdb.