fix: don't import schema file when in schema dump ci
If you're importing the same file as you're editing, it's not too useful. For example, one could edit the schema, add an entry to the bottom to the migrations table, and the ci will import it when running all migrations before generating the new dump. The new dump would therefore include what the schema file already had instead of the sql that the laravel migrations generated.