fix: filled request filter broken when fill approver soft deleted
Use the timestamp instead of the user id as it doesn't disappear when the user is (soft) deleted. The `whereDoesntHave` is the real issue here since it handles soft deletion, but since users are also deletable, handle all cases where it checks the field directly as well to instead use the timestamp.