fix: occasional tmdb ssl errors break `fetch:meta` command
Sometimes, there is an ssl error from tmdb (Once every few hundred thousand http requests). When this happens, because the jobs are processed synchronously to avoid dos'ing tmdb, it throws an exception and kills the command. Instead, allow the http requests to be retried, and skip the fetch if it still fails after 3 retries.