refactor: remove unused middlewares 3
I missed some more from the other times (#4744 and #4749). The json check in Authenticate.php has since been upstreamed into laravel, so we don't need to modify that explicitly anymore and can swap it out for laravel's middleware. The RedirectIfAuthenticated middleware now has an option to set the route to redirect to, so we can call that method directly instead of overwriting the method in the middleware. The TrimStrings middleware also now has an option to exclude certain attributes, so we can call that method directly instead of overwriting the method in the middleware.

TrustProxies.php is the only remaining Laravel middleware that is customized, but that can only be replaced with the laravel 11 app structure.