refactor: remove unused ticket/comment events/listeners/notifications
These aren't currently being used. If this functionality is to be completed, it should be easy enough to find the code through the git history again. However, UNIT3D doesn't really use events and it'd be better for most of the notification sending to be inline (except the stale one of course).