remove: chatbox connecting status
It causes a flash of content disappearing and reappearing. It looks far better for the old messages to remain on the screen for an extra half second then to hide the messages when switching chatbox tabs. It also looks better when first loading the home page to have no messages at first and have a blank chatbox window for half a second instead of the chatbox window disappearing. When this happens, the featured torrents flash in to fill up the empty space which looks glitchy.